Current Commit

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,7 +6,7 @@
  *
  * Return: Length.
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 	size_t len = 0;
 
@@ -15,4 +15,3 @@ size_t _strlen(const char *str)
 
 	return (len);
 }
-
